relocate example tests
