+++ cpp/examples/controller/CMakeLists.txt
@@ -5,3 +5,5 @@ target_link_libraries(example_control_lib
         eta_hsm_lib
         wise_enum
 )
+
+add_subdirectory(tests)
